[REFEAT]: CommentTopView 구현으로 ViewController 코드 경량화
