+++ Projects/Presentation/Challenge/Implementations/Member/Feed/FeedComment/Views/CommentPresentationModel.swift
@@ -1,0 +1,21 @@
+//
+//  CommentPresentationModel.swift
+//  Presentation
+//
+//  Created by jung on 12/17/24.
+//  Copyright © 2024 com.photi. All rights reserved.
+//
+
+import Foundation
+
+struct CommentPresentationModel: Hashable {
+  let id: String
+  let userName: String
+  let content: String
+  let isOwner: Bool
+  let updatedAt: Date
+  
+  func hash(into hasher: inout Hasher) {
+    hasher.combine(id)
+  }
+}
